fix: click on result field will call getSuggestedQuery from @testing-library/dom
