+++ src/components/Result.js
@@ -36,8 +36,7 @@ function Result({ result, dispatch }) {
       </div>
     );
   }
-  const { data, suggestion } = result.elements[0];
-
+  const { data, suggestion, target } = result.elements[0];
   return (
     <div className="flex flex-col w-full h-full overflow-hidden">
       <div className="flex-none pb-4 border-b">
@@ -56,6 +55,7 @@ function Result({ result, dispatch }) {
             suggestion={suggestion}
             activeMethod={result.expression?.method}
             dispatch={dispatch}
+            target={target}
           />
         </Scrollable>
       </div>
